+++ src/pool_thread/mod.rs
@@ -14,10 +14,10 @@ where
     P: PoolItem,
 {
     /// A unique id assigned to the pool thread
-    thread_id: usize,
+    thread_id: u64,
     /// Stores the channel on which requests will be received
     pool_thread_receiver: Receiver<SenderCouplet<P>>,
     /// This is a hash map that will hold the ownership of all pool items created in this
     /// pool thread keyed by their ids
-    pool_item_map: BTreeMap<usize, P>,
+    pool_item_map: BTreeMap<u64, P>,
 }
